% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinygenui-package.R
\docType{package}
\name{shinygenui-package}
\alias{shinygenui}
\alias{shinygenui-package}
\title{shinygenui: Generative UI for 'shiny'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Build interactive user interfaces for 'shiny' applications through a conversation with a large language model (LLM). Developers choose a set of reusable components, and the model arranges and updates those components as the user describes what they need. Each component's inputs are checked before it is shown, and the model supplies data rather than executable code. Applications can also save and replay the sequence of interface changes without contacting a model. For background on generative user interfaces, see Leviathan et al. (2026) \doi{10.48550/arXiv.2604.09577}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/shinygenui/}
  \item \url{https://github.com/nanxstats/shinygenui}
  \item Report bugs at \url{https://github.com/nanxstats/shinygenui/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID}) [copyright holder]
}

}
\keyword{internal}
