% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinygenui-package.R
\docType{package}
\name{shinygenui-package}
\alias{shinygenui}
\alias{shinygenui-package}
\title{shinygenui: Generative UI for 'shiny'}
\description{
What the package does (one paragraph, 2+ sentences).
}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID}) [copyright holder]
}

}
\keyword{internal}
